fix(core): changed align & pad type: uint8_t -> size_t

+++ include/oo_alloc/ArenaAllocator.hpp
@@ -13,11 +13,11 @@ public:
   ArenaAllocator();
   ~ArenaAllocator() override;
 
-  void* alloc(std::size_t size, std::uint8_t align) override;
+  void* alloc(std::size_t size, std::size_t align) override;
   void  free(void* ptr) override; // WARN: noop
   bool  init(std::size_t size) override;
   void clear() override;
-  void* realloc(void* ptr, std::size_t old_size, std::size_t new_size, std::uint8_t align) override;
+  void* realloc(void* ptr, std::size_t old_size, std::size_t new_size, std::size_t align) override;
 };
 
 }
